requirements.txt
certifi>=2021.5.30
qiskit-aer>=0.10.2
qiskit-terra>=0.19.2
amazon-braket-sdk>=1.18.0
